backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
argon2-cffi
pydantic-settings
python-multipart
libsql
email-validator
